+++ src/module/macros.ts
@@ -1,4 +1,5 @@
 import OD6S from "./config/config-od6s";
+import {MESSAGE_MODES} from "./hooks/chat-mode";
 
 /**
  * Create a Macro from an Item drop.
@@ -128,7 +129,7 @@ interface SimpleRollResult {
 async function runSimpleRoll(result: SimpleRollResult): Promise<void> {
     let wild = false;
     let rollString = "";
-    let rollMode = CONST.DICE_ROLL_MODES.PUBLIC;
+    let messageMode: string = MESSAGE_MODES.PUBLIC;
     let dice = result.dice;
     const pips = result.pips;
     const damageRoll = !!result.damageroll;
@@ -183,13 +184,13 @@ async function runSimpleRoll(result: SimpleRollResult): Promise<void> {
     }
 
     if (game.user.isGM && game.settings.get("nonex-ist-od6s", "hide-gm-rolls")) {
-        rollMode = CONST.DICE_ROLL_MODES.PRIVATE;
+        messageMode = MESSAGE_MODES.GM;
     }
     const rollMessage = await roll.toMessage({
         speaker: ChatMessage.getSpeaker(),
         flavor: label,
         flags: {"nonex-ist-od6s": flags},
-    }, {rollMode, create: true});
+    }, {messageMode, create: true});
 
     if (flags.wild === true && OD6S.wildDieOneDefault === 2 && OD6S.wildDieOneAuto === 0) {
         const replacementRoll = JSON.parse(JSON.stringify(rollMessage.rolls[0].toJSON()));